+++ test/test.js
@@ -1,8 +1,11 @@
-'use strict';
+import {join, dirname} from 'path';
+import {fileURLToPath} from 'url';
 
-const {join} = require('path');
+import sinon from 'sinon';
 
-const preassembledWorkerBoxes = require('../src/index.js');
+import preassembledWorkerBoxes from '../src/index.js';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
  * @param {string} file
@@ -13,8 +16,17 @@ function getFixturePath (file) {
 }
 
 describe('preassembledWorkerBoxes', function () {
-  it('Bakes in resources (with no imports)', async function () {
-    const {info, additionalManifestEntries} = await preassembledWorkerBoxes({
+  beforeEach(function () {
+    this.sinon = sinon.createSandbox();
+  });
+  afterEach(function () {
+    this.sinon.restore();
+  });
+
+  it('Bakes in resources (with no imports) and logFiles', async function () {
+    const {
+      info, additionalManifestEntries, logFiles
+    } = await preassembledWorkerBoxes({
       file: getFixturePath('fetches.js')
     });
     // console.log('info', info);
@@ -38,6 +50,26 @@ describe('preassembledWorkerBoxes', function () {
     expect(info.warnings).to.have.lengthOf(1);
     expect(info.warnings[0]).to.include('without revisioning info');
     expect(info.size).to.equal(0);
+
+    this.sinon.spy(console, 'log');
+    logFiles();
+    /* eslint-disable no-console -- Testing */
+    expect(console.log.calledWith(
+      'File count', 3
+    )).to.equal(true);
+    expect(console.log.calledWith(
+      'Combined size in bytes of all precached entries', 0
+    )).to.equal(true);
+    expect(console.log.calledWithMatch(
+      'Paths of precached entry files'
+    )).to.equal(true);
+
+    const arr = console.log.getCall(2).lastArg;
+    /* eslint-enable no-console -- Testing */
+    expect(arr[0]).to.have.string('/sw.js.map');
+    expect(arr[1]).to.have.string('/sw.js');
+    expect(arr[2]).to.have.string('/workbox-f905291c.js.map');
+    expect(arr[3]).to.have.string('/workbox-f905291c.js');
   });
 
   it('Bakes in resources (single import)', async function () {